cleanup(core): create a shared add-babel-input function
