+++ packages/react-native/package.json
@@ -28,6 +28,7 @@
     "@nrwl/detox": "file:../detox",
     "@nrwl/devkit": "file:../devkit",
     "@nrwl/jest": "file:../jest",
+    "@nrwl/js": "file:../js",
     "@nrwl/linter": "file:../linter",
     "@nrwl/react": "file:../react",
     "@nrwl/storybook": "file:../storybook",
