+++ packages/react-native/src/generators/init/init.ts
@@ -9,6 +9,8 @@ import {
 } from '@nrwl/devkit';
 import { Schema } from './schema';
 import { runTasksInSerial } from '@nrwl/workspace/src/utilities/run-tasks-in-serial';
+import { addBabelInputs } from '@nrwl/js/src/utils/add-babel-inputs';
+
 import { jestInitGenerator } from '@nrwl/jest';
 import { detoxInitGenerator } from '@nrwl/detox';
 import {
@@ -37,11 +39,10 @@ import {
 } from '../../utils/versions';
 
 import { addGitIgnoreEntry } from './lib/add-git-ignore-entry';
-import { initRootBabelConfig } from './lib/init-root-babel-config';
 
 export async function reactNativeInitGenerator(host: Tree, schema: Schema) {
   addGitIgnoreEntry(host);
-  initRootBabelConfig(host);
+  addBabelInputs(host);
 
   const tasks: GeneratorCallback[] = [];
