+++ packages/webpack/src/generators/init/init.ts
@@ -3,10 +3,7 @@ import {
   convertNxGenerator,
   formatFiles,
   GeneratorCallback,
-  readWorkspaceConfiguration,
   Tree,
-  updateWorkspaceConfiguration,
-  writeJson,
 } from '@nrwl/devkit';
 import { Schema } from './schema';
 import { swcCoreVersion } from '@nrwl/js/src/utils/versions';
@@ -15,12 +12,13 @@ import {
   swcLoaderVersion,
   tsLibVersion,
 } from '../../utils/versions';
+import { addBabelInputs } from '@nrwl/js/src/utils/add-babel-inputs';
 
 export async function webpackInitGenerator(tree: Tree, schema: Schema) {
   let task: GeneratorCallback;
 
   if (schema.compiler === 'babel') {
-    initRootBabelConfig(tree);
+    addBabelInputs(tree);
   }
 
   if (schema.compiler === 'swc') {
@@ -44,25 +42,6 @@ export async function webpackInitGenerator(tree: Tree, schema: Schema) {
   }
 
   return task;
-}
-
-function initRootBabelConfig(tree: Tree) {
-  if (tree.exists('/babel.config.json') || tree.exists('/babel.config.js')) {
-    return;
-  }
-
-  writeJson(tree, '/babel.config.json', {
-    babelrcRoots: ['*'], // Make sure .babelrc files other than root can be loaded in a monorepo
-  });
-
-  const workspaceConfiguration = readWorkspaceConfiguration(tree);
-
-  if (workspaceConfiguration.namedInputs?.sharedGlobals) {
-    workspaceConfiguration.namedInputs.sharedGlobals.push(
-      '{workspaceRoot}/babel.config.json'
-    );
-  }
-  updateWorkspaceConfiguration(tree, workspaceConfiguration);
 }
 
 export default webpackInitGenerator;
